Improve note tag handling
