+++ src/components/mainContainer/innerSideBar/notes/index.tsx
@@ -6,7 +6,7 @@ import useWindowSize from "@/hooks/useWindowSize"
 import { useNotesStore } from "@/stores/notes.store"
 import { sortAndFilterNotes } from "@/components/notes/utils"
 import { useLocalStorage } from "@uidotdev/usehooks"
-import Note from "./note"
+import Note from "@components/notes/components/NoteItem"
 import { validate as validateUUID } from "uuid"
 import { useNavigate } from "@tanstack/react-router"
 import useRouteParent from "@/hooks/useRouteParent"
