+++ src/components/mainContainer/topBar/notes.tsx
@@ -3,7 +3,7 @@ import { useNotesStore } from "@/stores/notes.store"
 import { CheckCircle2, Loader, MoreVertical, RefreshCwOff } from "lucide-react"
 import { showInputDialog } from "@/components/dialogs/input"
 import worker from "@/lib/worker"
-import ContextMenu from "../innerSideBar/notes/note/contextMenu"
+import ContextMenu from "@components/notes/components/NoteContextMenu"
 import useLoadingToast from "@/hooks/useLoadingToast"
 import useErrorToast from "@/hooks/useErrorToast"
 import { cn } from "@/lib/utils"
