+++ src/components/notes/components/NoteItem.tsx
@@ -4,7 +4,7 @@ import useRouteParent from "@/hooks/useRouteParent"
 import { Link } from "@tanstack/react-router"
 import { Archive, Trash, Text, ListChecks, Code, NotepadText, Pin, Heart, BookMarked } from "lucide-react"
 import { type Note as NoteType } from "@filen/sdk/dist/types/api/v3/notes"
-import ContextMenu from "./contextMenu"
+import ContextMenu from "./NoteContextMenu"
 import { simpleDate } from "@/utils"
 import ChatAvatar from "@/components/chatAvatar"
 import eventEmitter from "@/lib/eventEmitter"
